[Fix] - int64_t -> intmax_t

+++ libft/ft_intlen.c
@@ -6,11 +6,13 @@
 /*   By: zshanabe <zshanabe@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/26 19:55:56 by zshanabe          #+#    #+#             */
-/*   Updated: 2018/04/19 17:11:31 by zshanabe         ###   ########.fr       */
+/*   Updated: 2018/05/20 18:47:28 by zshanabe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_intlen(int n)
+#include "libft.h"
+
+int		ft_intlen(intmax_t n)
 {
 	int len;
 	int num;
